+++ src/ValueBuilders/CodeCaller.php
@@ -34,64 +34,27 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Exceptions/Traits
+ * @package   Exceptions/ValueBuilders
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
- * @link      http://code.ganbarodigital.com/php-file-system
+ * @link      http://code.ganbarodigital.com/php-exceptions
  */
 
-namespace GanbaroDigital\Exceptions;
+namespace GanbaroDigital\Exceptions\ValueBuilders;
 
-use PHPUnit_Framework_TestCase;
-use Exception;
-
-class ExceptionMessageDataTest_Target extends Exception
+class CodeCaller
 {
-    use ExceptionMessageData;
-}
-
-/**
- * @coversDefaultClass GanbaroDigital\Exceptions\ExceptionMessageData
- */
-class ExceptionMessageDataTest extends PHPUnit_Framework_TestCase
-{
-    /**
-     * @coversNone
-     */
-    public function testCanInstantiateTargetClass()
+    public static function fromBacktrace($backtrace)
     {
-        // ----------------------------------------------------------------
-        // perform the change
+        $retval = [null, null];
+        if (isset($backtrace[1]['class'])) {
+            $retval[0] = $backtrace[1]['class'];
+        }
+        if (isset($backtrace[1]['function'])) {
+            $retval[1] = $backtrace[1]['function'];
+        }
 
-        $obj = new ExceptionMessageDataTest_Target();
-
-        // ----------------------------------------------------------------
-        // test the results
-
-        $traits = class_uses($obj);
-        $this->assertTrue(in_array(ExceptionMessageData::class, $traits));
+        return $retval;
     }
-
-    /**
-     * @coversNothing
-     */
-    public function testHasExpectedTraitForBackwardsCompatibility()
-    {
-        // ----------------------------------------------------------------
-        // setup your test
-
-        $obj = new ExceptionMessageDataTest_Target();
-
-        // ----------------------------------------------------------------
-        // perform the change
-
-
-
-        // ----------------------------------------------------------------
-        // test the results
-
-        $this->assertTrue(method_exists($obj, 'getMessageData'));
-    }
-
 }