+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## develop branch
 
-Nothing yet
+Nothing yet.
+
+## 1.1.0 - Sat Jul 4 2015
+
+### New
+
+- CodeCaller value builder (moved from ganbarodigital/php-reflection)
+- ExceptionCaller trait
+- ExceptionMessageData trait (in the Traits folder)
+- UnsupportedType trait
 
 ## 1.0.0 - Sat Jun 27 2015